Remove base file, move code to the model file itself
